refactor: mobile gnb 내의 language select는 한 곳에서 관리하도록 (desktop용과 같이)
